Asset control.

+++ app/controller/RecordController.php
@@ -29,32 +29,40 @@ class RecordController extends Mind
 
     public function edit()
     {
-        if (
-            $this->do_have($this->TableName, $this->post['id'], 'id') and
-            !empty($this->post['name']) and
-            $this->is_phone($this->post['phone']) and
-            $this->is_email($this->post['email'])
-        ) {
-            $options = array(
-                'name' => $this->post['name'],
-                'phone' => $this->post['phone'],
-                'email' => $this->post['email'],
-                'updated_at' => $this->timestamp,
-            );
-            $this->update($this->TableName, $options, $this->post['id']);
-            $this->redirect();
+        if(isset($this->post['id'])){
+
+            if (
+                $this->do_have($this->TableName, $this->post['id'], 'id') and
+                !empty($this->post['name']) and
+                $this->is_phone($this->post['phone']) and
+                $this->is_email($this->post['email'])
+            ) {
+                $options = array(
+                    'name' => $this->post['name'],
+                    'phone' => $this->post['phone'],
+                    'email' => $this->post['email'],
+                    'updated_at' => $this->timestamp,
+                );
+                $this->update($this->TableName, $options, $this->post['id']);
+                $this->redirect();
+            }
+            
         }
     }
 
     public function remove()
     {
-        if (
-        $this->do_have($this->TableName, $this->post['id'], 'id')
-        ) {
-            if ($this->post['status'] == true) {
-                $this->delete($this->TableName, $this->post['id']);
-                $this->redirect();
+        if(isset($this->post['id'])){
+
+            if (
+            $this->do_have($this->TableName, $this->post['id'], 'id')
+            ) {
+                if ($this->post['status'] == true) {
+                    $this->delete($this->TableName, $this->post['id']);
+                    $this->redirect();
+                }
             }
+            
         }
     }
 }
